Tried to catch any OverflowException because the Part2 results wouldn't add upp when itterating over more rounds, but with no OverflowException luck... In the end I would expexct to change the code to hand over data between itterations but I would like to know when it failed with Part1 solution...

+++ Common/Helper.cs
@@ -23,7 +23,7 @@ namespace Common
             return number == 0;
         }
 
-        public static readonly Dictionary<char, Func<int, int, int>> Operators = new ()
+        public static readonly Dictionary<char, Func<long, long, long>> Operators = new ()
         {
             {'+', (x,y) => x+y},
             {'-', (x,y) => x-y},
